backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
python-multipart==0.0.9
jsonschema==4.23.0
